# openapi/openapi-ticket.json
{"openapi":"3.0.0","paths":{"/ticket/health":{"get":{"operationId":"HealthController_health","parameters":[],"responses":{"200":{"description":""}}}},"/ticket/admin":{"get":{"operationId":"TicketAdminController_getTickets","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeepTicketDto"}}}}}},"tags":["Ticket-Admin"],"security":[{"jwt":[]}]}},"/ticket/admin/{id}":{"get":{"operationId":"TicketAdminController_getTicketById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeepTicketDto"}}}}},"tags":["Ticket-Admin"],"security":[{"jwt":[]}]},"patch":{"operationId":"TicketAdminController_updateTicket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTicketDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDto"}}}}},"tags":["Ticket-Admin"],"security":[{"jwt":[]}]},"delete":{"operationId":"TicketAdminController_deleteTicket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Ticket-Admin"],"security":[{"jwt":[]}]}},"/ticket/admin/experience/{id}":{"get":{"operationId":"TicketAdminController_getTicketsForExperience","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketDto"}}}}}},"tags":["Ticket-Admin"],"security":[{"jwt":[]}]}},"/ticket/merchant":{"get":{"operationId":"TicketMerchantController_getTickets","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DeepTicketDto"}}}}}},"tags":["Ticket-Merchant"],"security":[{"jwt":[]}]},"post":{"operationId":"TicketMerchantController_createTicket","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTicketDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDto"}}}}},"tags":["Ticket-Merchant"],"security":[{"jwt":[]}]}},"/ticket/merchant/{id}":{"get":{"operationId":"TicketMerchantController_getTicketById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeepTicketDto"}}}}},"tags":["Ticket-Merchant"],"security":[{"jwt":[]}]},"patch":{"operationId":"TicketMerchantController_updateTicket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTicketDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TicketDto"}}}}},"tags":["Ticket-Merchant"],"security":[{"jwt":[]}]},"delete":{"operationId":"TicketMerchantController_deleteTicket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Ticket-Merchant"],"security":[{"jwt":[]}]}},"/ticket/merchant/experience/{id}":{"get":{"operationId":"TicketMerchantController_getTicketsForExperience","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketDto"}}}}}},"tags":["Ticket-Merchant"],"security":[{"jwt":[]}]}},"/ticket/{id}":{"get":{"operationId":"TicketController_getMerchant","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeepTicketDto"}}}}},"tags":["Ticket"]}},"/ticket/experience/{id}":{"get":{"operationId":"TicketController_getTicketsForExperience","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TicketDto"}}}}}},"tags":["Ticket"]}}},"info":{"title":"Ticketpond - ticket","description":"","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"cookie":{"type":"apiKey","in":"cookie","name":"jwt"}},"schemas":{"ExperienceDto":{"type":"object","properties":{"id":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"merchantId":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"name":{"type":"string","example":"Summer Festival"},"description":{"type":"string","example":"A festival for the summer"},"startDate":{"format":"date-time","type":"string","example":"2023-06-01T00:00:00.000Z"},"endDate":{"format":"date-time","type":"string","example":"2023-06-10T00:00:00.000Z"},"bannerImage":{"type":"string","example":"https://www.example.com/image.png"}},"required":["id","merchantId","name","description","startDate","endDate","bannerImage"]},"DeepTicketDto":{"type":"object","properties":{"id":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"name":{"type":"string","example":"One day ticket"},"description":{"type":"string","example":"This ticket is valid for one day"},"price":{"type":"number","example":123},"validFrom":{"format":"date-time","type":"string","example":"2023-06-01T00:00:00.000Z"},"validTo":{"format":"date-time","type":"string","example":"2023-06-02T00:00:00.000Z"},"experienceId":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"experience":{"$ref":"#/components/schemas/ExperienceDto"}},"required":["id","name","description","price","validFrom","validTo","experienceId","experience"]},"TicketDto":{"type":"object","properties":{"id":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"},"name":{"type":"string","example":"One day ticket"},"description":{"type":"string","example":"This ticket is valid for one day"},"price":{"type":"number","example":123},"validFrom":{"format":"date-time","type":"string","example":"2023-06-01T00:00:00.000Z"},"validTo":{"format":"date-time","type":"string","example":"2023-06-02T00:00:00.000Z"},"experienceId":{"type":"string","example":"550e8400-e29b-41d4-a716-446655440000"}},"required":["id","name","description","price","validFrom","validTo","experienceId"]},"UpdateTicketDto":{"type":"object","properties":{"name":{"type":"string","example":"One day ticket"},"description":{"type":"string","example":"This ticket is valid for one day"},"price":{"type":"number","example":123},"validFrom":{"format":"date-time","type":"string","example":"2023-06-01T00:00:00.000Z"},"validTo":{"format":"date-time","type":"string","example":"2023-06-02T00:00:00.000Z"}}},"CreateTicketDto":{"type":"object","properties":{"name":{"type":"string","example":"One day ticket"},"description":{"type":"string","example":"This ticket is valid for one day"},"price":{"type":"number","example":123},"validFrom":{"format":"date-time","type":"string","example":"2023-06-01T00:00:00.000Z"},"validTo":{"format":"date-time","type":"string","example":"2023-06-02T00:00:00.000Z"},"experienceId":{"type":"string"}},"required":["name","description","price","validFrom","validTo","experienceId"]}}}}